Move AutoQuant review helpers
