+++ examples/llm_ptq/hf_ptq.py
@@ -27,6 +27,8 @@ from accelerate.hooks import remove_hook_from_module
 from cast_mxfp4_to_nvfp4 import apply_to_model as apply_cast_mxfp4_to_nvfp4
 from cast_mxfp4_to_nvfp4 import force_weight_quantizers_static
 from example_utils import (
+    _get_auto_quantize_cost_excluded_patterns,
+    _get_auto_quantize_disabled_layers,
     build_quant_cfg,
     copy_custom_model_files,
     create_vlm_calibration_loop,
@@ -73,7 +75,7 @@ from modelopt.torch.export import (
 )
 from modelopt.torch.export.model_utils import get_language_model_from_vl, is_multimodal_model
 from modelopt.torch.quantization._auto_quantize_cost import EXCLUDED_MODULE_NAME_PATTERNS_KEY
-from modelopt.torch.quantization.config import _default_disabled_quantizer_cfg, need_calibration
+from modelopt.torch.quantization.config import need_calibration
 from modelopt.torch.quantization.plugins.accelerate import init_quantized_weights
 from modelopt.torch.quantization.utils import is_quantized
 from modelopt.torch.speculative.eagle.utils import (
@@ -157,59 +159,6 @@ def _canonical_qformat(name: str) -> str:
 
 
 mto.enable_huggingface_checkpointing()
-
-
-# TODO: Refactor into the config system.
-_QWEN36_AUTOQ_DISABLED_LAYERS = (
-    "*shared_expert_gate*",
-    "*linear_attn.in_proj_a*",
-    "*linear_attn.in_proj_b*",
-)
-_VLM_AUTOQ_DISABLED_LAYERS = ("*visual*", "*mtp*", "*vision_tower*")
-
-
-def _is_qwen_model(model) -> bool:
-    """Return True when model/config identifiers indicate a Qwen-family model."""
-    candidates = [type(model).__name__]
-    config = getattr(model, "config", None)
-    configs = [
-        config,
-        getattr(config, "text_config", None),
-        getattr(config, "language_config", None),
-    ]
-    for cfg in configs:
-        if cfg is None:
-            continue
-        candidates.append(type(cfg).__name__)
-        model_type = getattr(cfg, "model_type", None)
-        if model_type is not None:
-            candidates.append(str(model_type))
-        architectures = getattr(cfg, "architectures", ()) or ()
-        if isinstance(architectures, str):
-            architectures = (architectures,)
-        candidates.extend(str(architecture) for architecture in architectures)
-    return any("qwen" in candidate.lower() for candidate in candidates)
-
-
-def _get_auto_quantize_disabled_layers(model) -> list[str]:
-    """Return layer patterns that should be excluded from AutoQuantize search."""
-    disabled_layers = [
-        entry["quantizer_name"]
-        for entry in _default_disabled_quantizer_cfg
-        if "parent_class" not in entry and entry["quantizer_name"] != "*lm_head*"
-    ]
-    if _is_qwen_model(model):
-        disabled_layers.extend(p for p in _QWEN36_AUTOQ_DISABLED_LAYERS if p not in disabled_layers)
-    if is_multimodal_model(model):
-        disabled_layers.extend(p for p in _VLM_AUTOQ_DISABLED_LAYERS if p not in disabled_layers)
-    return disabled_layers
-
-
-def _get_auto_quantize_cost_excluded_patterns(model) -> list[str]:
-    """Return layer patterns excluded only from AutoQuantize cost accounting."""
-    if is_multimodal_model(model):
-        return list(_VLM_AUTOQ_DISABLED_LAYERS)
-    return []
 
 
 def extract_and_prepare_language_model_from_vl(full_model):
